+++ src/Plundi.Hammerfall.Core/Models/DamageHit.cs
@@ -2,7 +2,7 @@ namespace Plundi.Hammerfall.Core.Models;
 
 public record DamageHit
 {
-    public double Damage { get; init; }
+    public decimal Damage { get; init; }
     public bool IsRelative { get; init; }
-    public double Timing { get; init; }
+    public decimal Timing { get; init; }
 }
