+++ src/Plundi.Hammerfall.Core/Models/DamageRange.cs
@@ -2,8 +2,8 @@ namespace Plundi.Hammerfall.Core.Models;
 
 public record DamageRange
 {
-    public double Min { get; init; }
-    public double Max { get; init; }
-    public double MinDps { get; init; }
-    public double MaxDps { get; init; }
+    public decimal Min { get; init; }
+    public decimal Max { get; init; }
+    public decimal MinDps { get; init; }
+    public decimal MaxDps { get; init; }
 }
